refactor(app): move code to stop importing lw page
There were a couple helpers and types that lived in
app/src/pages/Desktop/Labware that were imported all over the place in
organisms. Let's give them a new home: local-resources

In the same way that resources/ provides app-specific wrappers around
robot api concerns, local-resources/ provides app-specific wrappers
around code and handling of other opentrons resources, like (for now)
labware.
