+++ app/src/organisms/Desktop/Labware/LabwareCard/__tests__/LabwareCard.test.tsx
@@ -3,14 +3,14 @@ import { screen } from '@testing-library/react'
 import { describe, it, vi, beforeEach } from 'vitest'
 import { renderWithProviders, nestedTextMatcher } from '/app/__testing-utils__'
 import { i18n } from '/app/i18n'
-import { useAllLabware } from '/app/pages/Desktop/Labware/hooks'
+import { useAllLabware } from '/app/local-resources/labware'
 import { mockDefinition } from '/app/redux/custom-labware/__fixtures__'
 import { CustomLabwareOverflowMenu } from '../CustomLabwareOverflowMenu'
 import { LabwareCard } from '..'
 
 import type * as OpentronsComponents from '@opentrons/components'
 
-vi.mock('/app/pages/Desktop/Labware/hooks')
+vi.mock('/app/local-resources/labware')
 vi.mock('../CustomLabwareOverflowMenu')
 
 vi.mock('@opentrons/components', async importOriginal => {
