+++ app/src/organisms/Desktop/Labware/LabwareDetails/WellSpacing.tsx
@@ -5,7 +5,7 @@ import { getSpacingDiagram } from '@opentrons/components'
 import { LabeledValue } from './StyledComponents/LabeledValue'
 import { ExpandingTitle } from './StyledComponents/ExpandingTitle'
 
-import type { LabwareWellGroupProperties } from '/app/pages/Desktop/Labware/types'
+import type { LabwareWellGroupProperties } from '/app/local-resources/labware'
 
 const toFixed = (n: number): string => round(n, 2).toFixed(2)
