+++ app/src/organisms/QuickTransferFlow/SelectSourceLabware.tsx
@@ -17,7 +17,7 @@ import { getCompatibleLabwareByCategory } from './utils'
 
 import type { LabwareDefinition2 } from '@opentrons/shared-data'
 import type { SmallButton } from '/app/atoms/buttons'
-import type { LabwareFilter } from '/app/pages/Desktop/Labware/types'
+import type { LabwareFilter } from '/app/local-resources/labware'
 import type {
   QuickTransferWizardState,
   QuickTransferWizardAction,
