--- app/src/pages/Desktop/Labware/helpers/getAllDefs.ts
@@ -1,6 +1,0 @@
-import { getAllDefinitions } from '@opentrons/shared-data'
-import type { LabwareDefinition2 } from '@opentrons/shared-data'
-
-export function getAllDefs(): LabwareDefinition2[] {
-  return Object.values(getAllDefinitions())
-}
