+++ app/src/pages/Desktop/Labware/index.tsx
@@ -35,16 +35,20 @@ import {
   ANALYTICS_OPEN_LABWARE_CREATOR_FROM_BOTTOM_OF_LABWARE_LIBRARY_LIST,
 } from '/app/redux/analytics'
 import { addCustomLabwareFileFromCreator } from '/app/redux/custom-labware'
-import { LabwareCard } from '/app/organisms/Desktop/LabwareCard'
-import { AddCustomLabwareSlideout } from '/app/organisms/Desktop/AddCustomLabwareSlideout'
-import { LabwareDetails } from '/app/organisms/Desktop/LabwareDetails'
+import { LabwareCard } from '/app/organisms/Desktop/Labware/LabwareCard'
+import { AddCustomLabwareSlideout } from '/app/organisms/Desktop/Labware/AddCustomLabwareSlideout'
+import { LabwareDetails } from '/app/organisms/Desktop/Labware/LabwareDetails'
 import { useToaster } from '/app/organisms/ToasterOven'
 import { useFeatureFlag } from '/app/redux/config'
-import { useAllLabware, useLabwareFailure, useNewLabwareName } from './hooks'
+import { useLabwareFailure, useNewLabwareName } from './hooks'
+import { useAllLabware } from '/app/local-resources/labware'
 
 import type { DropdownOption } from '@opentrons/components'
-import type { LabwareFilter, LabwareSort } from './types'
-import type { LabwareDefAndDate } from './hooks'
+import type {
+  LabwareFilter,
+  LabwareSort,
+  LabwareDefAndDate,
+} from '/app/local-resources/labware'
 
 const LABWARE_CREATOR_HREF = 'https://labware.opentrons.com/create/'
 const labwareDisplayCategoryFilters: LabwareFilter[] = [
