+++ app/src/redux/custom-labware/__fixtures__/index.ts
@@ -1,5 +1,5 @@
 import type { LabwareDefinition2 } from '@opentrons/shared-data'
-import type { LabwareWellGroupProperties } from '/app/pages/Desktop/Labware/types'
+import type { LabwareWellGroupProperties } from '/app/local-resources/labware'
 import type * as Types from '../types'
 
 export const mockDefinition: LabwareDefinition2 = {
